+++ lib/WorkflowCutandrun.groovy
@@ -36,6 +36,10 @@ class WorkflowCutandrun {
                 ucscGenomeWarn(log)
             }
         }
+
+        if (params.dt_calc_all_matrix) {
+            matrixWarn(log)
+        }
     }
 
     //
@@ -143,5 +147,16 @@ class WorkflowCutandrun {
         log.warn "=============================================================================\n" +
             "  No genome blacklist file specified, switching to dummy empty file...\n" +
             "==================================================================================="
+    }
+
+    //
+    // Print a warning if gen all plots are on
+    //
+    private static void matrixWarn(log) {
+        log.warn "==========================================================================================================\n" +
+            "  dt_calc_all_matrix is switched on which will calculate a deeptools matrix for all samples. \n" +
+            "  If you have a large sample count, this may affect pipeline performance and result in errors. \n" +
+            "  Set this option to false to disable this feature and only calculate deeptools heatmaps for single samples\n" +
+            "==============================================================================================================="
     }
 }